Created README.md
Today is mom's funeral service, and I am dreading all the people that will be there and not knowing how to behave and afraid of something like what happened in the hospital right before she died when I had been taking a new medication that gave me restless legs so bad and I guess reinforced everyone's idea that I was entirely annoyed with the topic of conversation because of some evil spirit or because of drugs that made me fidgety. It was the drugs that made me fidgety. They just came from a pharmacy instead of dealer.