Update README.md
Realized I was being too nice and included a little more purple. I mean truth. If you want tekhelet then you'll have to wait until my own cup is replaced with one that isn't full of holes and other people's siphon tubes who think that by destroying every bit of fun I could possibly have they are doing the Will of Almighty God saving the world from an evil sex crazed Jezzebel spirited monster. Get over yourselves and take a look in the mirror. You are the terrorists and not me. And my allegiance is not with anyone except Jesus Christ so do not even think for a second that I wouldn't throw anyone else under the bus in a heartbeat for a dirty look even.